Reactor to more easily shared utility
